+++ example/play_buffer_generator.dart
@@ -8,10 +8,10 @@ Future<void> main() async {
   final synthizer = Synthizer()..initialize();
   print('Synthizer initialized.');
   final ctx = synthizer.createContext();
-  print('Created context $ctx with gain ${ctx.gain}.');
+  print('Created context $ctx with gain ${ctx.gain.value}.');
   final buffer = Buffer.fromFile(synthizer, File('sound.wav'));
   final generator = ctx.createBufferGenerator(buffer: buffer);
-  print('Created generator $generator with gain ${generator.gain}.');
+  print('Created generator $generator with gain ${generator.gain.value}.');
   generator.gain.value = 0.5;
   ctx.createDirectSource().addGenerator(generator);
   await Future<void>.delayed(Duration(seconds: 2));